Move calculateUserLimits to utils
